[DEV-5291] Cleanup and prep for handling awards
